fix: satisfy pkg-file-count after cursor storage and provider splits
Move Cursor inference and provider-session detail code into subpackages,
consolidate cursorstorage store helpers into fewer files, and keep
backend-size/pkg-maint passing so Build, Lint, and API can succeed on CI.
