+++ pkg/internal/cursorstorage/database.go
@@ -63,3 +63,10 @@ type KeyValuePair struct {
 	Key   string
 	Value string
 }
+
+// Logging helpers ported from iksnae/cursor-session; disabled in-process for server use.
+
+func LogError(string, ...any)  {}
+func LogWarn(string, ...any)   {}
+func LogInfo(string, ...any)   {}
+func LogDebug(string, ...any)  {}
