--- pkg/internal/cursorstorage/logger.go
@@ -1,8 +1,0 @@
-package cursorstorage
-
-// Logging helpers ported from iksnae/cursor-session; disabled in-process for server use.
-
-func LogError(string, ...any)  {}
-func LogWarn(string, ...any)   {}
-func LogInfo(string, ...any)   {}
-func LogDebug(string, ...any)  {}
