+++ pkg/workers/provider/cursor/inference_test.go
@@ -1,4 +1,4 @@
-package provider
+package cursor
 
 import (
 	"encoding/json"
@@ -8,7 +8,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 )
 
-func TestParseCursorInferenceResult_Success(t *testing.T) {
+func TestParseInferenceResult_Success(t *testing.T) {
 	stdout := []byte(`{
 		"type": "result",
 		"subtype": "success",
@@ -26,9 +26,9 @@ func TestParseCursorInferenceResult_Success(t *testing.T) {
 		}
 	}`)
 
-	parsed, err := parseCursorInferenceResult(string(interfaces.ModelProviderCursor), stdout)
+	parsed, err := ParseInferenceResult(string(interfaces.ModelProviderCursor), stdout)
 	if err != nil {
-		t.Fatalf("parseCursorInferenceResult returned error: %v", err)
+		t.Fatalf("ParseInferenceResult returned error: %v", err)
 	}
 	if parsed.Content != "Done reviewing the repo." {
 		t.Fatalf("content = %q, want parsed result text", parsed.Content)
@@ -39,25 +39,25 @@ func TestParseCursorInferenceResult_Success(t *testing.T) {
 	if parsed.ProviderSession.Provider != string(interfaces.ModelProviderCursor) {
 		t.Fatalf("provider = %q, want cursor", parsed.ProviderSession.Provider)
 	}
-	if parsed.ProviderSession.Kind != providerSessionKindSessionID {
+	if parsed.ProviderSession.Kind != ProviderSessionKindSessionID {
 		t.Fatalf("kind = %q, want session_id", parsed.ProviderSession.Kind)
 	}
 	if parsed.ProviderSession.ID != "c6b62c6f-7ead-4fd6-9922-e952131177ff" {
 		t.Fatalf("session id = %q", parsed.ProviderSession.ID)
 	}
 
-	assertCursorResponseMetadata(t, parsed.ResponseMetadata, map[string]string{
-		cursorResponseMetadataRequestID:        "10e11780-df2f-45dc-a1ff-4540af32e9c0",
-		cursorResponseMetadataDurationMS:       "1234",
-		cursorResponseMetadataDurationAPIMS:    "1100",
-		cursorResponseMetadataInputTokens:      "1200",
-		cursorResponseMetadataOutputTokens:     "340",
-		cursorResponseMetadataCacheReadTokens:  "50",
-		cursorResponseMetadataCacheWriteTokens: "10",
+	assertResponseMetadata(t, parsed.ResponseMetadata, map[string]string{
+		ResponseMetadataRequestID:        "10e11780-df2f-45dc-a1ff-4540af32e9c0",
+		ResponseMetadataDurationMS:       "1234",
+		ResponseMetadataDurationAPIMS:    "1100",
+		ResponseMetadataInputTokens:      "1200",
+		ResponseMetadataOutputTokens:     "340",
+		ResponseMetadataCacheReadTokens:  "50",
+		ResponseMetadataCacheWriteTokens: "10",
 	})
 }
 
-func TestParseCursorInferenceResult_MissingSessionID(t *testing.T) {
+func TestParseInferenceResult_MissingSessionID(t *testing.T) {
 	stdout := []byte(`{
 		"type": "result",
 		"subtype": "success",
@@ -66,7 +66,7 @@ func TestParseCursorInferenceResult_MissingSessionID(t *testing.T) {
 		"session_id": ""
 	}`)
 
-	_, err := parseCursorInferenceResult(string(interfaces.ModelProviderCursor), stdout)
+	_, err := ParseInferenceResult(string(interfaces.ModelProviderCursor), stdout)
 	if err == nil {
 		t.Fatal("expected parse error for missing session_id")
 	}
@@ -75,8 +75,8 @@ func TestParseCursorInferenceResult_MissingSessionID(t *testing.T) {
 	}
 }
 
-func TestParseCursorInferenceResult_MalformedJSON(t *testing.T) {
-	_, err := parseCursorInferenceResult(string(interfaces.ModelProviderCursor), []byte(`{not json`))
+func TestParseInferenceResult_MalformedJSON(t *testing.T) {
+	_, err := ParseInferenceResult(string(interfaces.ModelProviderCursor), []byte(`{not json`))
 	if err == nil {
 		t.Fatal("expected parse error for malformed JSON")
 	}
@@ -85,10 +85,10 @@ func TestParseCursorInferenceResult_MalformedJSON(t *testing.T) {
 	}
 }
 
-func TestParseCursorInferenceResult_UnexpectedType(t *testing.T) {
+func TestParseInferenceResult_UnexpectedType(t *testing.T) {
 	stdout := []byte(`{"type":"assistant","subtype":"success","result":"hi","session_id":"sess-1"}`)
 
-	_, err := parseCursorInferenceResult(string(interfaces.ModelProviderCursor), stdout)
+	_, err := ParseInferenceResult(string(interfaces.ModelProviderCursor), stdout)
 	if err == nil {
 		t.Fatal("expected parse error for unexpected type")
 	}
@@ -97,7 +97,7 @@ func TestParseCursorInferenceResult_UnexpectedType(t *testing.T) {
 	}
 }
 
-func TestParseCursorInferenceResult_ErrorSubtype(t *testing.T) {
+func TestParseInferenceResult_ErrorSubtype(t *testing.T) {
 	stdout := []byte(`{
 		"type": "result",
 		"subtype": "error",
@@ -106,7 +106,7 @@ func TestParseCursorInferenceResult_ErrorSubtype(t *testing.T) {
 		"session_id": "sess-1"
 	}`)
 
-	_, err := parseCursorInferenceResult(string(interfaces.ModelProviderCursor), stdout)
+	_, err := ParseInferenceResult(string(interfaces.ModelProviderCursor), stdout)
 	if err == nil {
 		t.Fatal("expected parse error for error subtype")
 	}
@@ -115,56 +115,47 @@ func TestParseCursorInferenceResult_ErrorSubtype(t *testing.T) {
 	}
 }
 
-func cursorSuccessStdoutJSON(result, sessionID string) []byte {
-	if result == "" {
-		result = "Done. COMPLETE"
-	}
-	if sessionID == "" {
-		sessionID = "cursor-test-session"
-	}
-	payload := cursorResultPayload{
-		Type:      cursorResultTypeResult,
-		Subtype:   cursorResultSubtypeSuccess,
-		SessionID: sessionID,
-		Result:    result,
-	}
-	encoded, err := json.Marshal(payload)
-	if err != nil {
-		panic(err)
-	}
-	return encoded
-}
-
 func TestBoundedCommandOutputExcerpt_TruncatesWhenOverLimit(t *testing.T) {
 	const limit = 8
 	long := []byte("0123456789abcdef")
-	got := boundedCommandOutputExcerpt(long, limit)
+	got := BoundedCommandOutputExcerpt(long, limit)
 	want := "01234567..."
 	if got != want {
 		t.Fatalf("excerpt = %q, want %q", got, want)
 	}
 }
 
-func TestWithCursorCommandOutputExcerpts_AttachesBoundedStdoutAndStderr(t *testing.T) {
-	stdout := []byte(strings.Repeat("a", cursorCommandOutputExcerptLimit+10))
+func TestWithCommandOutputExcerpts_AttachesBoundedStdoutAndStderr(t *testing.T) {
+	stdout := []byte(strings.Repeat("a", CommandOutputExcerptLimit+10))
 	stderr := []byte("rate limited\n")
-	diagnostics := withCursorCommandOutputExcerpts(nil, stdout, stderr)
+	diagnostics := WithCommandOutputExcerpts(nil, stdout, stderr)
 	if diagnostics == nil || diagnostics.Provider == nil {
 		t.Fatal("expected provider diagnostics with excerpts")
 	}
-	if got := diagnostics.Provider.ResponseMetadata[cursorResponseMetadataStdoutExcerpt]; len(got) != cursorCommandOutputExcerptLimit+3 {
-		t.Fatalf("stdout excerpt len = %d, want %d with ellipsis", len(got), cursorCommandOutputExcerptLimit+3)
+	if got := diagnostics.Provider.ResponseMetadata[ResponseMetadataStdoutExcerpt]; len(got) != CommandOutputExcerptLimit+3 {
+		t.Fatalf("stdout excerpt len = %d, want %d with ellipsis", len(got), CommandOutputExcerptLimit+3)
 	}
-	if got := diagnostics.Provider.ResponseMetadata[cursorResponseMetadataStderrExcerpt]; got != "rate limited" {
+	if got := diagnostics.Provider.ResponseMetadata[ResponseMetadataStderrExcerpt]; got != "rate limited" {
 		t.Fatalf("stderr excerpt = %q, want rate limited", got)
 	}
 }
 
-func assertCursorResponseMetadata(t *testing.T, metadata map[string]string, want map[string]string) {
+func assertResponseMetadata(t *testing.T, metadata map[string]string, want map[string]string) {
 	t.Helper()
 	for key, wantValue := range want {
 		if got := metadata[key]; got != wantValue {
 			t.Fatalf("response metadata[%q] = %q, want %q", key, got, wantValue)
 		}
+	}
+}
+
+func TestSuccessStdoutJSON(t *testing.T) {
+	encoded := SuccessStdoutJSON("hello", "sess-1")
+	var payload resultPayload
+	if err := json.Unmarshal(encoded, &payload); err != nil {
+		t.Fatalf("unmarshal: %v", err)
+	}
+	if payload.Result != "hello" || payload.SessionID != "sess-1" {
+		t.Fatalf("payload = %#v", payload)
 	}
 }
