+++ pkg/workers/provider/inference_provider.go
@@ -13,6 +13,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/logging"
 	"github.com/portpowered/infinite-you/pkg/workcontent/materialize"
+	cursorpkg "github.com/portpowered/infinite-you/pkg/workers/provider/cursor"
 	workerprocess "github.com/portpowered/infinite-you/pkg/workers/process"
 )
 
@@ -335,10 +336,56 @@ func cursorFailureLogFields(req interfaces.RunnerExecutionRequest, cursorProvide
 		return fields
 	}
 	fields = append(fields,
-		"stdout_preview", boundedCommandOutputExcerpt(result.Stdout, cursorCommandOutputLogPreviewLimit),
-		"stderr_preview", boundedCommandOutputExcerpt(result.Stderr, cursorCommandOutputLogPreviewLimit),
+		"stdout_preview", cursorpkg.BoundedCommandOutputExcerpt(result.Stdout, cursorpkg.CommandOutputLogPreviewLimit),
+		"stderr_preview", cursorpkg.BoundedCommandOutputExcerpt(result.Stderr, cursorpkg.CommandOutputLogPreviewLimit),
 	)
 	return fields
+}
+
+func cursorInferenceFailureDiagnostics(
+	cursorProvider bool,
+	commandDiagnostics *interfaces.WorkDiagnostics,
+	result CommandResult,
+) *interfaces.WorkDiagnostics {
+	if !cursorProvider {
+		return commandDiagnostics
+	}
+	return cursorpkg.WithCommandOutputExcerpts(commandDiagnostics, result.Stdout, result.Stderr)
+}
+
+func (p *ScriptWrapProvider) completeCursorInference(
+	req interfaces.RunnerExecutionRequest,
+	result CommandResult,
+	commandDiagnostics *interfaces.WorkDiagnostics,
+	logger logging.Logger,
+) (interfaces.InferenceResponse, error) {
+	parsed, parseErr := cursorpkg.ParseInferenceResult(req.ModelProvider, result.Stdout)
+	if parseErr != nil {
+		logger.Error("inferencer: cursor JSON parse failed",
+			cursorFailureLogFields(req, true, result,
+				"error", parseErr.Message)...)
+		failureDiagnostics := cursorpkg.WithCommandOutputExcerpts(commandDiagnostics, result.Stdout, result.Stderr)
+		return interfaces.InferenceResponse{}, newProviderErrorWithDiagnostics(
+			parseErr.Type,
+			parseErr.Message,
+			parseErr.Cause,
+			effectiveProviderSession(req, result),
+			failureDiagnostics,
+		)
+	}
+	diagnostics := cursorpkg.WithResponseMetadata(commandDiagnostics, parsed.ResponseMetadata)
+	logger.Debug("inference results:",
+		workLogFields(req.Dispatch.Execution, "output", parsed.Content)...)
+	logger.Info("inferencer: request completed",
+		workLogFields(req.Dispatch.Execution,
+			"dispatcher", string(req.ModelProvider),
+			"output_len", len(parsed.Content),
+			"session_id", parsed.ProviderSession.ID)...)
+	return interfaces.InferenceResponse{
+		Content:         parsed.Content,
+		ProviderSession: parsed.ProviderSession,
+		Diagnostics:     diagnostics,
+	}, nil
 }
 
 func formatProviderCommandFailure(provider string, result CommandResult, err error) string {
